Auto Price Manager
The AUTOPRICEMANAGER system was designed by Sellers for Sellers to
enable you to get the most sales at the best price all the time. We have
left out just as much as we have put in, we did this to make it quick
and simple and achieve the end goal of winning you the sale without
having to over-discount your Listings.